chore(velocity_tracking): move velocity tracking to device finder
The Velocity tracking on the controllers was originally done in
Controller Events, but this is not the best place for it as velocity
retrieval is not really an event. The Velocity and Angular Velocity
retrieval methods have now been moved to the Device Finder script.

The Controller SDK Update routine has also been moved out of
Controller Events and into the Tracked Controller script as this script
is always available in the scene.

An addition of a Tracked Headset script has also been added that runs
an Update routine for the headset within the SDK mimicking the
functionality of the Tracked Controller update.

The methods in the Controller Events script are still in place but have
been deprecated (and removed from the documentation) and now all they
do is pass through the call to the Device Finder.